Refactor: Extract getting video url from watch_media
